+++ view/src/__mocks/handlers/index.ts
@@ -1,4 +1,5 @@
 import { transactionHandlers } from './transactionHandlers.ts';
 import { balanceHandlers } from './balanceHandlers.ts';
+import websocketHandlers from './websocketHandlers.ts';
 
-export const handlers = [...transactionHandlers, ...balanceHandlers];
+export const handlers = [...transactionHandlers, ...balanceHandlers, ...websocketHandlers];
